fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
psycopg2-binary
python-dotenv